Modified the user routes to respond with jwt tokens and set them in cookies
