+++ backend/package.json
@@ -19,6 +19,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
     "joi": "^17.12.2",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.2.0",
     "nodemon": "^3.1.0"
   }
