+++ client/src/components/Logout.jsx
@@ -9,7 +9,9 @@ const LogoutForm = ({ setIsUserLoggedIn }) => {
   const handleLogout = async () => {
     try {
       document.cookie =
-        "username=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;";
+        "user=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;";
+        document.cookie =
+          "token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;";
       setIsUserLoggedIn(false);
       console.log("User logged out successfully");
       toast.success("LogOut Successful!");
