+++ client/src/components/UpdateForm.jsx
@@ -27,7 +27,7 @@ const UpdateForm = ({ currentMeme, setInitiateUpdate }) => {
 
       console.log(filteredData);
       const response = await axios.patch(
-        `https://frantic-smock-lion.cyclic.app/patch/${mood_category}/${meme_id}`,
+        `http://localhost:8080/meme/patch/${mood_category}/${meme_id}`,
         filteredData
       );
